WebspotSih.md
"Background: There are more than 1000 universities in India and these universities are offering thousands of undergraduate, post-Graduate and doctorate courses. The students applying for these universities have to do a lot of comparison for choosing the university/course on the basis of various criteria’s. A PAN India university information bank is required for accessing course details, comparison of courses, facilities and faculty, admission, tracking student mobility, viewing results, evaluation and certification, verification of certificates, joint online courses (MOOCs). This will help both student and faculty communities for self improvement through comparison and emulations. Lack of information is leading to self glorification and inbreeding among University and College student and faculty members. Summary: The universities in India provide various courses on different disciplines and the students have to compare these courses on the basis of various parameters. The courses range from Bachelor, Masters and P.hd courses. The integration of these details on a single platform will enhance the transparency and remove difficulty being faced by the students due to lack of information. Objective: : To develop an integrated platform to provide the information of the courses offered by Indian universities"